requirements.txt
customtkinter >= 5.1.2
pillow >= 9.5.0
